+++ src/main/java/jscover/util/DateTime.java
@@ -1,0 +1,7 @@
+package jscover.util;
+
+import java.util.Date;
+
+public interface DateTime {
+    Date getDate();
+}