+++ src/main/java/jscover/util/DateTimeImpl.java
@@ -1,0 +1,9 @@
+package jscover.util;
+
+import java.util.Date;
+
+public class DateTimeImpl implements DateTime {
+    public Date getDate() {
+        return new Date();
+    }
+}